@@ -6,21 +6,21 @@
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@mui/material": "^5.15.19",
-    "axios": "^0.21.4",
+    "axios": "^1.7.2",
     "chess.js": "^0.13.4",
-    "react": "^17.0.2",
-    "react-chessboard": "^1.3.1",
+    "react": "^18.3.1",
+    "react-chessboard": "^4.6.0",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.23.1",
     "socket.io-client": "^4.7.5"
   },
   "devDependencies": {
-    "autoprefixer": "^10.2.5",
+    "autoprefixer": "^10.4.19",
     "cross-env": "^7.0.3",
-    "gh-pages": "^3.2.3",
-    "postcss": "^8.2.15",
-    "react-scripts": "^4.0.3",
-    "tailwindcss": "^2.2.4"
+    "gh-pages": "^6.1.1",
+    "postcss": "^8.4.38",
+    "react-scripts": "^5.0.1",
+    "tailwindcss": "^3.4.4"
   },
   "scripts": {
     "start": "react-scripts start",
