@@ -5,16 +5,39 @@ import Login from './pages/Login';
 import Register from './pages/Register';
 import Games from './pages/Games';
 import Game from './pages/Game';
+import ProtectedRoute from './components/ProtectedRoute';
 
 function App() {
   return (
-    <Router basename="/">
+    <Router>
       <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<Login />} />
         <Route path="/login" element={<Login />} />
         <Route path="/register" element={<Register />} />
-        <Route path="/games" element={<Games />} />
-        <Route path="/game/:gameId" element={<Game />} /> {/* Actualizado */}
+        <Route
+          path="/home"
+          element={
+            <ProtectedRoute>
+              <Home />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/games"
+          element={
+            <ProtectedRoute>
+              <Games />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/game/:gameId"
+          element={
+            <ProtectedRoute>
+              <Game />
+            </ProtectedRoute>
+          }
+        />
       </Routes>
     </Router>
   );
